+++ src/main/java/com/mageddo/dnsproxyserver/docker/dataprovider/DockerNetworkFacadeDefault.java
@@ -8,7 +8,7 @@ import com.github.dockerjava.api.model.Network;
 import com.mageddo.commons.lang.Objects;
 import com.mageddo.dnsproxyserver.docker.application.Containers;
 import com.mageddo.dnsproxyserver.docker.domain.NetworkConnectionStatus;
-import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.DockerNetworkService;
+import com.mageddo.dnsproxyserver.server.dns.solver.docker.Network.Name;
 import com.mageddo.net.Networks;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -142,7 +142,7 @@ public class DockerNetworkFacadeDefault implements DockerNetworkFacade {
       .withStatusFilter(Containers.RUNNING_STATUS_LIST)
       .exec()
       .stream()
-      .filter(it -> Boolean.FALSE.equals(DockerNetworkService.isHostNetwork(it)))
+      .filter(it -> Boolean.FALSE.equals(isHostNetwork(it)))
       .filter(it -> !Containers.containsNetworkName(it, networkName))
       .filter(p)
       .forEach(container -> this.connect(networkName, container.getId()))
@@ -152,6 +152,15 @@ public class DockerNetworkFacadeDefault implements DockerNetworkFacade {
   @Override
   public boolean exists(String networkId) {
     return this.findById(networkId) != null;
+  }
+
+  private static Boolean isHostNetwork(Container container) {
+    final var config = container.getHostConfig();
+    if (config == null) {
+      return null;
+    }
+    final var networkMode = config.getNetworkMode();
+    return Name.HOST.equalTo(networkMode);
   }
 
 }
